+++ portal.ps1
@@ -41,7 +41,7 @@
     PS C:\> .\portal.ps1 -Unattended        # Silent: export HTML and exit
 
 .NOTES
-    Version : 3.6
+    Version : 5.0
 
 #>