+++ pyre.ps1
@@ -47,7 +47,7 @@
     PS C:\> .\pyre.ps1 -Unattended        # Silent: export HTML and exit
 
 .NOTES
-    Version : 3.6
+    Version : 5.0
 
 #>