+++ runepress.ps1
@@ -34,7 +34,7 @@
     PS C:\> .\runepress.ps1 -Unattended        # Silent mode — auto-selects first INF, skips printer config
 
 .NOTES
-    Version : 3.6
+    Version : 5.0
 
 #>