+++ shade.ps1
@@ -35,7 +35,7 @@
         Enable-PSRemoting -Force
 
 .NOTES
-    Version : 3.6
+    Version : 5.0
 
     Remote-Compatible Tools
     ─────────────────────────────────────────────────────────────────