+++ talon.ps1
@@ -42,7 +42,7 @@
     PS C:\> .\talon.ps1 -Unattended        # Silent: export HTML and exit
 
 .NOTES
-    Version : 3.6
+    Version : 5.0
 
 #>