+++ totem.ps1
@@ -37,7 +37,7 @@
     PS C:\> .\totem.ps1 -Unattended        # Silent: export HTML and exit
 
 .NOTES
-    Version : 3.6
+    Version : 5.0
 
 #>